Ask user before deleting type, method, field

+++ ILSpy/AsmEditor/Field/FieldDefCommands.cs
@@ -123,6 +123,10 @@ namespace ICSharpCode.ILSpy.AsmEditor.Field
 			if (!CanExecute(nodes))
 				return;
 
+			var res = MainWindow.Instance.ShowIgnorableMessageBox("delete def", "There could be code in this assembly or in another assembly that references this field. Are you sure you want to delete the field?", System.Windows.MessageBoxButton.YesNo);
+			if (res != null && res != MsgBoxButton.OK)
+				return;
+
 			var fieldNodes = nodes.Select(a => (FieldTreeNode)a).ToArray();
 			UndoCommandManager.Instance.Add(new DeleteFieldDefCommand(fieldNodes));
 		}
