+++ ILSpy/AsmEditor/Types/TypeDefCommands.cs
@@ -134,6 +134,10 @@ namespace ICSharpCode.ILSpy.AsmEditor.Types
 			if (!CanExecute(nodes))
 				return;
 
+			var res = MainWindow.Instance.ShowIgnorableMessageBox("delete def", "There could be code in this assembly or in another assembly that references this type. Are you sure you want to delete the type?", System.Windows.MessageBoxButton.YesNo);
+			if (res != null && res != MsgBoxButton.OK)
+				return;
+
 			var typeNodes = FilterOutGlobalTypes(nodes).Select(a => (TypeTreeNode)a).ToArray();
 			UndoCommandManager.Instance.Add(new DeleteTypeDefCommand(typeNodes));
 		}
